requirements.txt
opencv-python==4.10.0.84
Pillow==10.4.0
python-dotenv==1.0.1
imageio==2.36.1
numpy==2.1.1